Move support methods into TypeAwareGitShard class
