+++ peyotl/git_storage/git_shard.py
@@ -4,7 +4,9 @@ import os
 import codecs
 import anyjson
 from threading import Lock
-from peyotl.utility import get_logger
+from peyotl.utility import get_logger, \
+                           write_to_filepath
+from peyotl.utility.input_output import read_as_json, write_as_json
 
 class FailedShardCreationError(ValueError):
     def __init__(self, message):
@@ -243,6 +245,33 @@ class TypeAwareGitShard(GitShard):
         ga = self.create_git_action()
         return ga.get_changed_docs(ancestral_commit_sha, doc_ids_to_check=doc_ids_to_check)
         #TODO:git-action-edits
+
+    def _read_master_branch_resource(self, fn, is_json=False):
+        '''This will force the current branch to master! '''
+        with self._master_branch_repo_lock:
+            ga = self._create_git_action_for_global_resource()
+            with ga.lock():
+                ga.checkout_master()
+                if os.path.exists(fn):
+                    if is_json:
+                        return read_as_json(fn)
+                    with codecs.open(fn, 'rU', encoding='utf-8') as f:
+                        ret = f.read()
+                    return ret
+                return None
+    def _write_master_branch_resource(self, content, fn, commit_msg, is_json=False):
+        '''This will force the current branch to master! '''
+        #TODO: we might want this to push, but currently it is only called in contexts in which
+        # we are about to push anyway (document creation)
+        with self._master_branch_repo_lock:
+            ga = self._create_git_action_for_global_resource()
+            with ga.lock():
+                ga.checkout_master()
+                if is_json:
+                    write_as_json(content, fn)
+                else:
+                    write_to_filepath(content, fn)
+                ga._add_and_commit(fn, self._infrastructure_commit_author, commit_msg)
 
 def _invert_dict_list_val(d):
     o = {}
