+++ peyotl/phylesystem/phylesystem_shard.py
@@ -3,13 +3,11 @@ import re
 import json
 import codecs
 from threading import Lock
-from peyotl.utility import get_config_setting_kwargs, \
-                           write_to_filepath
+from peyotl.utility import get_config_setting_kwargs
 from peyotl.git_storage.git_shard import GitShard, \
                                          TypeAwareGitShard, \
                                          FailedShardCreationError, \
                                          _invert_dict_list_val
-from peyotl.utility.input_output import read_as_json, write_as_json
 
 #_LOG = get_logger(__name__)
 #class PhylesystemShardBase(object):
@@ -278,33 +276,6 @@ class PhylesystemShard(TypeAwareGitShard):
                                            commit_msg=content,
                                            is_json=False)
         return c
-
-    def _read_master_branch_resource(self, fn, is_json=False):
-        '''This will force the current branch to master! '''
-        with self._master_branch_repo_lock:
-            ga = self._create_git_action_for_global_resource()
-            with ga.lock():
-                ga.checkout_master()
-                if os.path.exists(fn):
-                    if is_json:
-                        return read_as_json(fn)
-                    with codecs.open(fn, 'rU', encoding='utf-8') as f:
-                        ret = f.read()
-                    return ret
-                return None
-    def _write_master_branch_resource(self, content, fn, commit_msg, is_json=False):
-        '''This will force the current branch to master! '''
-        #TODO: we might want this to push, but currently it is only called in contexts in which
-        # we are about to push any way (study creation)
-        with self._master_branch_repo_lock:
-            ga = self._create_git_action_for_global_resource()
-            with ga.lock():
-                ga.checkout_master()
-                if is_json:
-                    write_as_json(content, fn)
-                else:
-                    write_to_filepath(content, fn)
-                ga._add_and_commit(fn, self._infrastructure_commit_author, commit_msg)
 
     def _diagnose_prefixes(self):
         '''Returns a set of all of the prefixes seen in the main document dir
